+++ app/Http/Controllers/ProductController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Product;
 use Exception;
 use Illuminate\Http\Request;
@@ -14,10 +15,10 @@ class ProductController extends Controller
      */
     public function index(Request $request)
     {
-        try{
+        try {
             $products = Product::with("category")->byTerm($request->term)->byCategory($request->category)->paginateOrNot($request->paginate, $request->per_page);
             return response()->json($products, 200);
-        }catch(Exception $e){
+        } catch (Exception $e) {
             return response()->json([
                 "success" => false,
                 "message" => $e->getMessage()
@@ -34,28 +35,28 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        try{
+        try {
 
-            $product=new Product();
-            $product->name=$request->name;
-            $product->barcode=$request->barcode;
-            $product->price=$request->price;
-            $product->presentation_quantity=0;
-            $product->presentation=$request->presentation;
-            $product->stock=$request->stock;
-            $product->min_stock=$request->min_stock;
-            $product->max_stock=$request->max_stock;
-            $product->brand=$request->brand;
-            $product->category_id=$request->category_id;
-            $product->unit=$request->unit;
-            $product->created_by=$request->user()->id;
-            $product->description=$request->description;
+            $product = new Product();
+            $product->name = $request->name;
+            $product->barcode = $request->barcode;
+            $product->price = $request->price;
+            $product->presentation_quantity = 0;
+            $product->presentation = $request->presentation;
+            $product->stock = $request->stock;
+            $product->min_stock = $request->min_stock;
+            $product->max_stock = $request->max_stock;
+            $product->brand = $request->brand;
+            $product->category_id = $request->category_id;
+            $product->unit = $request->unit;
+            $product->created_by = $request->user()->id;
+            $product->description = $request->description;
             $product->save();
             return response()->json([
                 "success" => true,
                 "message" => "Product created successfully"
             ], 200);
-        }catch(Exception $e){
+        } catch (Exception $e) {
             return response()->json([
                 "success" => false,
                 "message" => $e->getMessage()
@@ -71,11 +72,10 @@ class ProductController extends Controller
      */
     public function show($id)
     {
-        try{
-            $product=Product::findOrFail($id);
-            return response()->json(["success"=>true,"data"=>$product],200);
-
-        }catch(Exception $e){
+        try {
+            $product = Product::findOrFail($id);
+            return response()->json(["success" => true, "data" => $product], 200);
+        } catch (Exception $e) {
             return response()->json([
                 "success" => false,
                 "message" => $e->getMessage()
@@ -83,7 +83,7 @@ class ProductController extends Controller
         }
     }
 
-  
+
 
     /**
      * Update the specified resource in storage.
@@ -94,26 +94,26 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        try{
-            $product=Product::find($id);
-            $product->name=$request->name;
-            $product->barcode=$request->barcode;
-            $product->price=$request->price;
-            $product->presentation_quantity=0;
-            $product->presentation=$request->presentation;
-            $product->stock=$request->stock;
-            $product->min_stock=$request->min_stock;
-            $product->max_stock=$request->max_stock;
-            $product->brand=$request->brand;
-            $product->category_id=$request->category_id;
-            $product->unit=$request->unit;
-            $product->description=$request->description;
+        try {
+            $product = Product::find($id);
+            $product->name = $request->name;
+            $product->barcode = $request->barcode;
+            $product->price = $request->price;
+            $product->presentation_quantity = 0;
+            $product->presentation = $request->presentation;
+            $product->stock = $request->stock;
+            $product->min_stock = $request->min_stock;
+            $product->max_stock = $request->max_stock;
+            $product->brand = $request->brand;
+            $product->category_id = $request->category_id;
+            $product->unit = $request->unit;
+            $product->description = $request->description;
             $product->save();
             return response()->json([
                 "success" => true,
                 "message" => "Product updated successfully"
             ], 200);
-        }catch(Exception $e){
+        } catch (Exception $e) {
             return response()->json([
                 "success" => false,
                 "message" => $e->getMessage()
@@ -129,11 +129,14 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        try{
-            $product=Product::findOrFail($id);
+        try {
+            $product = Product::findOrFail($id);
+            if ($product->transfer_details()->count() > 0) {
+                throw new Exception("El producto no se puede eliminar porque esta asociado a una transferencia");
+            }
             $product->delete();
-            return response()->json(["success"=>true,"message"=>"Product deleted successfully"],200);
-        }catch(Exception $e){
+            return response()->json(["success" => true, "message" => "Product deleted successfully"], 200);
+        } catch (Exception $e) {
             return response()->json([
                 "success" => false,
                 "message" => $e->getMessage()
